requirements.txt
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.10.0
python-telegram-bot~=13.11